Project1 U.S. Employment Visas Analysis EDA
As the most developed country in the world, the United States attracts workers from all over the world, and countless people hope to get a job there. The Immigration and Nationality Act (INA) of the U.S. permits foreign workers to come to the United States to work On either a temporary or permanent basis. But at the same time, when ensuring that U.S. employers obtain qualified talents, they also need to protect the employment rights and status of U.S. citizens, which requires relevant departments (such as customs, etc.) to identify each foreign employee, Give approval or denial.

The data set contains many variables, including the applicant's education, work experience, salary, and approval. The objective of this study is we want to know the demographic of the applicants, and which characteristics of employees and employers are more likely to be granted visas, and we will perform the EDA and analysis of the data to answer this question.